integrations/cloudflare-haystack/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloudflare-haystack"
dynamic = ["version"]
description = 'Use Cloudflare AI products with Haystack'
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []
authors = [
  { name = "deepset GmbH", email = "info@deepset.ai" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "haystack-ai",
  "cloudflare-python",
]

[project.urls]
Documentation = "https://github.com/deepset GmbH/cloudflare-haystack#readme"
Issues = "https://github.com/deepset GmbH/cloudflare-haystack/issues"
Source = "https://github.com/deepset GmbH/cloudflare-haystack"

[tool.hatch.version]
path = "src/cloudflare_haystack/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/cloudflare_haystack tests}"

[tool.coverage.run]
source_pkgs = ["cloudflare_haystack", "tests"]
branch = true
parallel = true
omit = [
  "src/cloudflare_haystack/__about__.py",
]

[tool.coverage.paths]
cloudflare_haystack = ["src/cloudflare_haystack", "*/cloudflare-haystack/src/cloudflare_haystack"]
tests = ["tests", "*/cloudflare-haystack/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
